backend/requirements.txt
fastapi[all]==0.110.0
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.6.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.4
httpx==0.26.0
